% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyColorBrewer.R
\name{adjust.color.number}
\alias{adjust.color.number}
\title{Expand colors into the number you wanted}
\usage{
adjust.color.number(color, n)
}
\arguments{
\item{color}{The color vector}

\item{n}{The number of colors you wanted}
}
\value{
A list of colors you wanted.
}
\description{
Expand colors into the number you wanted
}
\examples{
adjust.color.number(color = c("red","green", "green"), n= 10)
}
