% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyColorBrewer.R
\name{brewer.colors}
\alias{brewer.colors}
\title{Get colors from RColorBrewer package}
\usage{
brewer.colors(name, n = NULL)
}
\arguments{
\item{name}{The names of color set. It must be one of the colors in \pkg{RColorBrewer} package. We suggest using the
following colors for heatmap: RdYlBu, RdBu, YlGnBu, PRGn.}

\item{n}{Number of colors you want. We suggest you set it by default. Default: NULL, means using the raw color elements.}
}
\value{
A character vector of colors.
}
\description{
Get colors from RColorBrewer package
}
\examples{
brewer.colors(name="RdBu")
}
