% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyColorBrewer.R
\name{browse.material.colors}
\alias{browse.material.colors}
\title{Browse material color information}
\usage{
browse.material.colors()
}
\value{
a data.frame with color name and element number.
}
\description{
Browse material color information
}
\examples{
browse.material.colors()
}
