% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyColorBrewer.R
\name{display.color.all}
\alias{display.color.all}
\title{Display all the colors}
\usage{
display.color.all(type = c("material.set", "material"), brightness = 3L)
}
\arguments{
\item{type}{The color set to display}

\item{brightness}{The brightness selected to display. Default: 3L, should between 1 and 10.}
}
\value{
A figure
}
\description{
Display all the colors
}
\examples{
display.color.all(type = "material.set")
}
