% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyColorBrewer.R
\name{display.color}
\alias{display.color}
\title{Display colors}
\usage{
display.color(colors, color_name = "")
}
\arguments{
\item{colors}{A vector of colors. Must be character.}

\item{color_name}{The name of colors.}
}
\value{
An image.
}
\description{
Display colors
}
\examples{
display.color(colors = c("#90CAF9", "#80DEEA", "#A5D6A7", "#FFF59D", "#FFCC80", "#EF9A9A", "#CE93D8", "#9FA8DA"))
}
