% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyColorBrewer.R
\name{material.color.set}
\alias{material.color.set}
\title{Get gradient colors}
\usage{
material.color.set(
  name = c("Developmental.basic", "Developmental.BuRd", "Developmental.full",
    "heatmap.BuRd", "heatmap.BuYlRd", "heatmap.GnPR", "heatmap.BuGnYl", "heatmap.BuPu",
    "heatmap.GnBuPu", "Rainbow", "Discrete"),
  n = NULL,
  brightness = 3L
)
}
\arguments{
\item{name}{The names of gradient color set. Developmental means gradient colors for sequential development. This is
designed for developmental process, for example, stem cell differentiation. We'll keep updating the beautiful colors.
Default: Developmental.basic.}

\item{n}{Number of colors you want. We suggest you set it by default. Default: NULL, means using the raw color elements.}

\item{brightness}{The brightness of colors, from 1 to 10. Default: 3}
}
\value{
A character vector of colors.
}
\description{
Get gradient colors
}
\examples{
color_sequential <- material.color.set("Developmental.basic")
display.color(color_sequential, "Developmental.basic")

}
