% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyColorBrewer.R
\name{material.colors}
\alias{material.colors}
\title{Get material colors}
\usage{
material.colors(name, brightness = 3L, mat_out = FALSE)
}
\arguments{
\item{name}{The name(s) of material colors. To browser all the names of material colors, use `browse.material.colors()`.}

\item{brightness}{The brightness of colors, from 1 to 10. Default: 3}

\item{mat_out}{Output a matrix?}
}
\value{
A vector of colors
}
\description{
Get material colors
}
\examples{
material.colors(name = "Red",brightness = 1:6)
}
