% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{simpletheme}
\alias{simpletheme}
\title{ggplot theme style}
\usage{
simpletheme(
  text_size = 6,
  text_color = "#000000",
  line_color = "#000000",
  line_size = 0.4672897,
  axis.text.x.angle = 0,
  axis.text.x.hjust = 0.5,
  axis.text.x.vjust = 0.5,
  ...
)
}
\arguments{
\item{text_size}{The text size}

\item{text_color}{Text color for axis. Default: black.}

\item{line_color}{Text color for axis. Default: black.}

\item{line_size}{The line size}

\item{axis.text.x.angle}{The angle of text in x axis.}

\item{axis.text.x.hjust}{Horizontal adjustment for x axis.}

\item{axis.text.x.vjust}{Vertical adjustment for x axis.}

\item{...}{others parameters.}
}
\value{
A ggplot theme.
}
\description{
ggplot theme style
}
\examples{
simpletheme()
}
