% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{theme_basic}
\alias{theme_basic}
\title{The basic themes of tidyplot}
\usage{
theme_basic(
  text.size = 6,
  line.size = 0.4672897,
  border.color = "#000000",
  axis.line.x.color = "#000000",
  axis.line.y.color = "#000000",
  panel.grid.major.x.color = "#e0e0e0",
  panel.grid.major.y.color = "#e0e0e0",
  axis.text.x.angle = 0,
  axis.text.x.hjust = 0.5,
  axis.text.x.vjust = 0
)
}
\arguments{
\item{text.size}{The size of text in the figure. Default: 6 pt. The Cell, Nature and Science journals
required the size of text should between 5 to 7 pt.}

\item{line.size}{The size of lines in the figure. Default: 1 pt. The Cell, Nature and Science journals
required the minimum size of lines should be greater than 0.3 pt.}

\item{border.color}{The color of borders of figures. Default: black.}

\item{axis.line.x.color}{The color of x axis. Default, black, not used when border color is black.}

\item{axis.line.y.color}{The color of y axis. Default, black, not used when border color is black.}

\item{panel.grid.major.x.color}{The color of major grid lines at x panel, Default: gray.}

\item{panel.grid.major.y.color}{The color of major grid lines at y panel, Default: gray.}

\item{axis.text.x.angle}{The angle of text at x axis.}

\item{axis.text.x.hjust}{The adjustment of text at x axis. Used when angle is not 0.}

\item{axis.text.x.vjust}{The adjustment of text at x axis. Used when angle is not 0.}
}
\value{
A theme object to ggplot2.
}
\description{
The basic themes of tidyplot
}
\examples{
library(ggplot2)
p0 <- ggplot2::ggplot(data = mpg, aes(x = trans, fill = factor(cyl))) + geom_bar(colour = 1, binwidth = 0.2)
p0 + theme_basic()
}
