% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{theme_slope}
\alias{theme_slope}
\title{The slope themes of tidyplot}
\usage{
theme_slope(
  axis.text.x.angle = 45,
  axis.text.x.hjust = 1,
  axis.text.x.vjust = 1,
  ...
)
}
\arguments{
\item{axis.text.x.angle}{The angle of text at x axis. Default:45.}

\item{axis.text.x.hjust}{The adjustment of text at x axis. Default: 1.}

\item{axis.text.x.vjust}{The adjustment of text at x axis. Default:1.}

\item{...}{Other parameters passed to 'tidytheme_basic()'.}
}
\value{
A theme object to ggplot2.
}
\description{
The slope themes of tidyplot
}
\examples{
library(ggplot2)
p0 <- ggplot(data = mpg, aes(x = trans, fill = factor(cyl))) + geom_bar(colour = 1, binwidth = 0.2)
p0 + theme_slope()
}
