% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{tidytheme_clean}
\alias{tidytheme_clean}
\title{The clean themes of tidyplot}
\usage{
tidytheme_clean(
  border.color = NA,
  axis.line.x.color = "#000000",
  axis.line.y.color = "#000000",
  panel.grid.major.x.color = NA,
  panel.grid.major.y.color = NA,
  ...
)
}
\arguments{
\item{border.color}{The color of borders of figures. Default: NA.}

\item{axis.line.x.color}{The color of x axis. Default: black.}

\item{axis.line.y.color}{The color of y axis. Default: black.}

\item{panel.grid.major.x.color}{The color of major grid lines at x panel, Default: NA.}

\item{panel.grid.major.y.color}{The color of major grid lines at y panel, Default: NA.}

\item{...}{Other parameters passed to 'tidytheme_basic()'.}
}
\value{
A theme object to ggplot2.
}
\description{
The clean themes of tidyplot
}
\examples{
library(ggplot2)
p0 <- ggplot(data = mpg, aes(x = trans, fill = factor(cyl))) + geom_bar(colour = 1, binwidth = 0.2)
p0 + tidytheme_clean()
}
