% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyColorBrewer.R
\name{material.colors}
\alias{material.colors}
\title{Get material colors}
\usage{
material.colors(name, n = 10)
}
\arguments{
\item{name}{The names of material colors.}

\item{n}{Number of colors you want to get. Default: 10, all color elements.}
}
\value{
A vector of colors
}
\description{
Get material colors
}
\examples{
material.colors(name = "Red",n = 6)
}
